修改弹窗式信息提示

+++ assets/js/common/islogin.js
@@ -15,13 +15,32 @@ var params = {
   code:"",
   imgCode:"",
   ucode:"",
+  channel:""
 }
 var stringParams = JSON.stringify(params,userReplacer).replace(/\"/g, "").replace(/\:/g, '=').replace(/\,/g, '&').replace(/\{/g, '').replace(/\}/g, '')
 params.sign = md5(stringParams)
-$.ajax({
+if(window.localStorage.getItem('token') == '' || window.localStorage.getItem('token') == null){
+  Swal.fire({
+      text: i18n.get('firstLogin'),
+      showCancelButton: false,
+      allowOutsideClick:false,
+      allowEscapeKey:false,
+      reverseButtons:true,
+      width:'16em',
+      confirmButtonColor: '#94cb82',
+      confirmButtonText: i18n.get('confirm'),
+  }).then((result) => {
+      if (result.isConfirmed) {
+          var url = window.location.href
+          window.localStorage.setItem('returnurl',url)
+          window.location.href = '../login/login.html'
+      }
+  })
+}else{
+  $.ajax({
     type:"POST",
     url:proURL+"/userinfo",
-    headers: {  
+    headers: {
       "channel" : "cloudasr"
     },
     data:params,
@@ -29,22 +48,22 @@ $.ajax({
       if(data.errorcode ==1024){
         $('.my_fav_list_a').css('pointer-events','none')
         window.localStorage.clear()
-        // Swal.fire({
-        //     text: i18n.get('firstLogin'),
-        //     showCancelButton: false,
-        //     allowOutsideClick:false,
-        //     allowEscapeKey:false,
-        //     reverseButtons:true,
-        //     width:'16em', 
-        //     confirmButtonColor: '#94cb82',
-        //     confirmButtonText: i18n.get('confirm'),
-        // }).then((result)=>{
-        //     if (result.isConfirmed) {
+        Swal.fire({
+            text: i18n.get('firstLogin'),
+            showCancelButton: false,
+            allowOutsideClick:false,
+            allowEscapeKey:false,
+            reverseButtons:true,
+            width:'16em', 
+            confirmButtonColor: '#94cb82',
+            confirmButtonText: i18n.get('confirm'),
+        }).then((result)=>{
+            if (result.isConfirmed) {
               var url = window.location.href
               window.localStorage.setItem('returnurl',url)
               window.location.href = '../login/login.html'
-        //     }
-        // })
+            }
+        })
       }else{
         $('.my_fav_list_a').css('pointer-events','auto')
       }
@@ -54,3 +73,4 @@ $.ajax({
         $('.login-reg').css("display","block")
     }
   })
+}
