+++ assets/js/forum/questiondetail.js
@@ -190,24 +190,24 @@ $(function () {
                                                 history.go(0)
                                             } else {
                                                 localStorage.clear();
-                                                // Swal.fire({
-                                                //     text: i18n.get('logTimeOut'),
-                                                //     showCancelButton: true,
-                                                //     allowOutsideClick:false,
-                                                //     allowEscapeKey:false,
-                                                //     reverseButtons:true,
-                                                //     width:'16em',
-                                                //     confirmButtonColor: '#94cb82',
-                                                //     cancelButtonColor: '#d33',
-                                                //     confirmButtonText: i18n.get('confirm'),
-                                                //     cancelButtonText:i18n.get('cancel')
-                                                // }).then((result) => {
-                                                //     if (result.isConfirmed) {
+                                                Swal.fire({
+                                                    text: i18n.get('logTimeOut'),
+                                                    showCancelButton: true,
+                                                    allowOutsideClick:false,
+                                                    allowEscapeKey:false,
+                                                    reverseButtons:true,
+                                                    width:'16em',
+                                                    confirmButtonColor: '#94cb82',
+                                                    cancelButtonColor: '#d33',
+                                                    confirmButtonText: i18n.get('confirm'),
+                                                    cancelButtonText:i18n.get('cancel')
+                                                }).then((result) => {
+                                                    if (result.isConfirmed) {
                                                         var url = window.location.href
                                                         window.localStorage.setItem('returnurl',url)
                                                         window.location.href = '../login/login.html'
-                                                //     }
-                                                // })
+                                                    }
+                                                })
                                                 
                                             }
                                         }
@@ -258,24 +258,24 @@ $(function () {
                                                 history.go(0)
                                             } else {
                                                 localStorage.clear();
-                                                // Swal.fire({
-                                                //     text: i18n.get('logTimeOut'),
-                                                //     showCancelButton: true,
-                                                //     allowOutsideClick:false,
-                                                //     allowEscapeKey:false,
-                                                //     reverseButtons:true,
-                                                //     width:'16em',
-                                                //     confirmButtonColor: '#94cb82',
-                                                //     cancelButtonColor: '#d33',
-                                                //     confirmButtonText: i18n.get('confirm'),
-                                                //     cancelButtonText:i18n.get('cancel')
-                                                // }).then((result) => {
-                                                //     if (result.isConfirmed) {
+                                                Swal.fire({
+                                                    text: i18n.get('logTimeOut'),
+                                                    showCancelButton: true,
+                                                    allowOutsideClick:false,
+                                                    allowEscapeKey:false,
+                                                    reverseButtons:true,
+                                                    width:'16em',
+                                                    confirmButtonColor: '#94cb82',
+                                                    cancelButtonColor: '#d33',
+                                                    confirmButtonText: i18n.get('confirm'),
+                                                    cancelButtonText:i18n.get('cancel')
+                                                }).then((result) => {
+                                                    if (result.isConfirmed) {
                                                         var url = window.location.href
                                                         window.localStorage.setItem('returnurl',url)
                                                         window.location.href = '../login/login.html'
-                                                //     }
-                                                // })
+                                                    }
+                                                })
                                             }
 
                                         }
@@ -292,24 +292,24 @@ $(function () {
                     }
                 } else {
                     localStorage.clear();
-                    // Swal.fire({
-                    //     text: i18n.get('logTimeOut'),
-                    //     showCancelButton: true,
-                    //     allowOutsideClick:false,
-                    //     allowEscapeKey:false,
-                    //     reverseButtons:true,
-                    //     width:'16em',
-                    //     confirmButtonColor: '#94cb82',
-                    //     cancelButtonColor: '#d33',
-                    //     confirmButtonText: i18n.get('confirm'),
-                    //     cancelButtonText:i18n.get('cancel')
-                    // }).then((result) => {
-                    //     if (result.isConfirmed) {
+                    Swal.fire({
+                        text: i18n.get('logTimeOut'),
+                        showCancelButton: true,
+                        allowOutsideClick:false,
+                        allowEscapeKey:false,
+                        reverseButtons:true,
+                        width:'16em',
+                        confirmButtonColor: '#94cb82',
+                        cancelButtonColor: '#d33',
+                        confirmButtonText: i18n.get('confirm'),
+                        cancelButtonText:i18n.get('cancel')
+                    }).then((result) => {
+                        if (result.isConfirmed) {
                             var url = window.location.href
                             window.localStorage.setItem('returnurl',url)
                             window.location.href = '../login/login.html'
-                    //     }
-                    // })
+                        }
+                    })
                 }
             },
             error: function (err) {
@@ -363,24 +363,24 @@ $(function () {
                                 history.go(0)
                             } else {
                                 localStorage.clear();
-                                // Swal.fire({
-                                //     text: i18n.get('logTimeOut'),
-                                //     showCancelButton: true,
-                                //     allowOutsideClick:false,
-                                //     allowEscapeKey:false,
-                                //     reverseButtons:true,
-                                //     width:'16em',
-                                //     confirmButtonColor: '#94cb82',
-                                //     cancelButtonColor: '#d33',
-                                //     confirmButtonText: i18n.get('confirm'),
-                                //     cancelButtonText:i18n.get('cancel')
-                                // }).then((result) => {
-                                //     if (result.isConfirmed) {
+                                Swal.fire({
+                                    text: i18n.get('logTimeOut'),
+                                    showCancelButton: true,
+                                    allowOutsideClick:false,
+                                    allowEscapeKey:false,
+                                    reverseButtons:true,
+                                    width:'16em',
+                                    confirmButtonColor: '#94cb82',
+                                    cancelButtonColor: '#d33',
+                                    confirmButtonText: i18n.get('confirm'),
+                                    cancelButtonText:i18n.get('cancel')
+                                }).then((result) => {
+                                    if (result.isConfirmed) {
                                         var url = window.location.href
                                         window.localStorage.setItem('returnurl',url)
                                         window.location.href = '../login/login.html'
-                                //     }
-                                // })
+                                    }
+                                })
                             }
 
                         },
@@ -499,24 +499,24 @@ $(function () {
                     }
             } else {
                 localStorage.clear();
-                // Swal.fire({
-                //     text: i18n.get('logTimeOut'),
-                //     showCancelButton: true,
-                //     allowOutsideClick:false,
-                //     allowEscapeKey:false,
-                //     reverseButtons:true,
-                //     width:'16em',
-                //     confirmButtonColor: '#94cb82',
-                //     cancelButtonColor: '#d33',
-                //     confirmButtonText: i18n.get('confirm'),
-                //     cancelButtonText:i18n.get('cancel')
-                // }).then((result) => {
-                //     if (result.isConfirmed) {
+                Swal.fire({
+                    text: i18n.get('logTimeOut'),
+                    showCancelButton: true,
+                    allowOutsideClick:false,
+                    allowEscapeKey:false,
+                    reverseButtons:true,
+                    width:'16em',
+                    confirmButtonColor: '#94cb82',
+                    cancelButtonColor: '#d33',
+                    confirmButtonText: i18n.get('confirm'),
+                    cancelButtonText:i18n.get('cancel')
+                }).then((result) => {
+                    if (result.isConfirmed) {
                         var url = window.location.href
                         window.localStorage.setItem('returnurl',url)
                         window.location.href = '../login/login.html'
-                //     }
-                // })
+                    }
+                })
             }
         },
         error: function (err) {
@@ -582,24 +582,24 @@ $(function () {
                     content()
                 } else {
                     localStorage.clear();
-                    // Swal.fire({
-                    //     text: i18n.get('logTimeOut'),
-                    //     showCancelButton: true,
-                    //     allowOutsideClick:false,
-                    //     allowEscapeKey:false,
-                    //     reverseButtons:true,
-                    //     width:'16em',
-                    //     confirmButtonColor: '#94cb82',
-                    //     cancelButtonColor: '#d33',
-                    //     confirmButtonText: i18n.get('confirm'),
-                    //     cancelButtonText:i18n.get('cancel')
-                    // }).then((result) => {
-                    //     if (result.isConfirmed) {
+                    Swal.fire({
+                        text: i18n.get('logTimeOut'),
+                        showCancelButton: true,
+                        allowOutsideClick:false,
+                        allowEscapeKey:false,
+                        reverseButtons:true,
+                        width:'16em',
+                        confirmButtonColor: '#94cb82',
+                        cancelButtonColor: '#d33',
+                        confirmButtonText: i18n.get('confirm'),
+                        cancelButtonText:i18n.get('cancel')
+                    }).then((result) => {
+                        if (result.isConfirmed) {
                             var url = window.location.href
                             window.localStorage.setItem('returnurl',url)
                             window.location.href = '../login/login.html'
-                    //     }
-                    // })
+                        }
+                    })
                 }
             },
             error: function (err) {
